Add FontRef::from_face function.
This allows FontRefs to be already created from pre-parsed faces. This can be useful to avoid having to reparse a face if we already have one, e.g. if we want to do shaping with rustybuzz, we should already have a parsed face and parsing it a second time is wasteful.